test: extract public functions
